rafactor: separates components logic
