+++ .babelrc.js
@@ -17,4 +17,6 @@ const plugins = [
   ],
 ];
 
-module.exports = { plugins };
+const presets = [["@babel/preset-react"]];
+
+module.exports = { plugins, presets };
