+++ package.json
@@ -49,10 +49,12 @@
     ]
   },
   "devDependencies": {
+    "@babel/preset-react": "^7.12.7",
     "babel-plugin-styled-components": "^1.12.0",
     "babel-plugin-transform-imports": "^2.0.0",
     "customize-cra": "^1.0.0",
     "react-app-rewired": "^2.1.6",
+    "react-styleguidist": "^11.1.5",
     "stylelint": "^13.8.0",
     "stylelint-config-rational-order": "^0.1.2",
     "stylelint-config-recommended": "^3.0.0",
