+++ styleguide.config.js
@@ -1,0 +1,7 @@
+const path = require("path");
+
+module.exports = {
+  styleguideComponents: {
+    Wrapper: path.join(__dirname, "src/redux/ReduxProvider"),
+  },
+};
